[UnitTest1]DrawGameBoard - Test 5
Successfully draws both a 3x3 and 4x4 game board game with the given number of rows and columns, each cell is populated with a value from an ordered array. The template is as described in the requirements. The values in the cells are left in as a template ready for player markers.